% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_functions_unconfined.R
\name{check_aquifer_depleted}
\alias{check_aquifer_depleted}
\title{Check if unconfined aquifer is fully depleted}
\usage{
check_aquifer_depleted(qs, qf, params, treaty)
}
\arguments{
\item{qs}{pumping from player S}

\item{qf}{pumping from player F}

\item{params}{parameter data.frame with single row}

\item{treaty}{boolean value that determines the context for evaluation (e.g., for Nash should be \code{F}, for First Best should be \code{T})}
}
\value{
Returns boolean value, TRUE if the aquifer has been fully depleted for some amount of pumping.
}
\description{
Is pumping enough to fully deplete the aquifer for either player?
}
\keyword{internal}
