% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_params.R
\name{check_game_dynamics}
\alias{check_game_dynamics}
\title{Check game dynamics}
\usage{
check_game_dynamics(params, text_results = TRUE, aquifer_type = NULL)
}
\arguments{
\item{params}{game parameters}

\item{text_results}{logical value determining if results are returned as text or a tibble}

\item{aquifer_type}{"confined" or "unconfined", or \code{NULL} -- in which case \code{check_params} will determine the type}
}
\value{
Either (1) a text description of the dynamics (if \code{text_results==TRUE}) or (2) a table
containing the following values:
\itemize{
\item: \code{ds_max, df_max}: The maximum depth for each player, if both players pump at their demand (\code{Qs, Qf}).
\item: \code{ds_threshold, df_threshold}: The depth for each player where the cost of pumping equals the cost of alternative supply (\code{Qi=B*di}).
}
}
\description{
Check the basic game dynamics, particularly in cases where zRange=0 and the first best
and nash outcomes are the same.
}
\details{
In a number of situations, the genevois game does nothing interesting: both players pump at their
entire demand, the first best is equivalent to the nash equilibrium, and there are no
advantages or disadvantages to signing a treaty. This function helps us better understand
why this is the case.

For any interesting dynamics to arise from the game, the First Best has to be different (better)
than the Nash Equilibrium. This occurs when players forgo pumping and supply their water from another
source, which can only occur if the other source is less expensive than pumping. In other words,
pumping for at least one of the players has to be great enough that the cost of pumping is greater
than the cost of alternative supply. If this is not the case, then the players cannot do any better
working together than they could on their own.
}
\examples{
library(magrittr)

# Table output
check_game_dynamics(example_params_unconfined,text_results = FALSE)
example_params_unconfined \%>\% dplyr::mutate(Qs=1,Qf=1) \%>\% check_game_dynamics(text_results = FALSE)

# Text output
check_game_dynamics(example_params_unconfined) \%>\% cat()
example_params_unconfined \%>\% dplyr::mutate(Qs=1,Qf=1) \%>\% check_game_dynamics() \%>\% cat()
}
