% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_params.R
\name{check_params}
\alias{check_params}
\title{Check for valid parameters}
\usage{
check_params(params)
}
\arguments{
\item{params}{\code{data.frame} or \code{list} of parameters}
}
\value{
Returns the aquifer type, depending on whether Dxx is specified ("confined") or PHIxx is specified ("unconfined")
}
\description{
Check for valid parameters
}
\details{
To evaluate the game, a \code{data.frame} or \code{list} of paramers must be supplied to
\code{evaluate_treaty} or \code{evaluate_treaty_cases}. Generally speaking, these parameters
should be non-negative. The parameters must include, for all scenarios:
\describe{
\item{Demand}{\code{Qf, Qs}}
\item{Unit cost of alternative supply}{\code{p0f, p0s}}
\item{Unit cost of pumping}{\code{B}}
\item{Unit cost of recharge, Swiss}{\code{crs}}
\item{Volume of recharge, with (T) and without (N) a treaty}{\code{rmT, rmN}}
\item{Trust between players}{\code{gs, gf}}
\item{Cost of the treaty}{\code{es, ef}}
}
And additional parameters depending on the type of aquifer. For confined aquifers:
\describe{
\item{Drawdown relationships}{\code{Dff, Dss, Dsf, Dfs}}
\item{Groundwater depth without pumping}{\code{d0s, d0f}}
\item{Recharge relationship with a treaty}{\code{DsrT, DfrT}}
\item{Recharge relationship without a treaty}{\code{DsrN, DfrN}}
}
And for unconfined aquifers:
\describe{
\item{Drawdown relationships}{\code{PHIff, PHIss, PHIsf, PHIfs}}
\item{Groundwater depth without pumping}{\code{dBs, dBf, h0s, h0f}}
\item{Recharge relationship with a treaty}{\code{PHIsrT, PHIfrT}}
\item{Recharge relationship without a treaty}{\code{PHIsrN, PHIfrN}}
}
}
\examples{
library(genevoisgame)
check_params(example_params_confined)
check_params(example_params_unconfined)
\dontrun{check_params(example_params_confined[,-ncol(example_params_confined):-1])}

params <- data.frame(example_params_confined,gs=c(0.2,0.8))
check_params(params)
\dontrun{check_params(params[,-ncol(example_params_confined):-1])}
}
