% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_evaluate.R
\name{evaluate_treaty_cases}
\alias{evaluate_treaty_cases}
\title{Evaluate multiple treaty scenarios}
\usage{
evaluate_treaty_cases(params_df, return_criteria = "qp")
}
\arguments{
\item{params_df}{Data.frame of parameters with each row sent to \code{evaluate_treaty()}.}

\item{return_criteria}{Character string containing letters that indicate output variables. See details.}
}
\value{
Returns a \code{tibble} containing z-values needed for the treaty
and whether or not there is a treaty (i.e., if zRange > 0), as well as output specified by \code{return_criteria}.
}
\description{
Evaluate whether or not the treaty will be made under multiple scenarios
}
\details{
Evaluate the treaty given multiple combinations of social, economic,
and geophysical parameters.
This function takes a data.frame of parameters,
evaluates each row to see if a treaty is signed,
and returns a tibble with the results and original params.

The parameter \code{return_criteria} can contain the following letters:
\itemize{
\item p - will return only parameters different from default. Otherwise all parameters returned
\item a - return all parameters (i.e., it's redundant to include a AND p)
\item u - return utilities of each player
\item d - return depth to water table for each player
}
}
\examples{
library(genevoisgame)
params_df <- genevoisgame::default_params
evaluate_treaty_cases(df)
evaluate_treaty_cases(default_params,'')
}
