% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_evaluate.R
\name{evaluate_treaty_depths}
\alias{evaluate_treaty_depths}
\title{Evaluate treaty depths}
\usage{
evaluate_treaty_depths(params, q_vals, aquifer_type)
}
\description{
Evaluate water table depth given (single) treaty parameters
}
