% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_evaluate.R
\name{evaluate_treaty_utility}
\alias{evaluate_treaty_utility}
\title{Evaluate treaty utility}
\usage{
evaluate_treaty_utility(params, q_vals, aquifer_type)
}
\description{
Evaluate utility given (single) treaty parameters
}
