% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\docType{data}
\name{example_params_confined}
\alias{example_params_confined}
\title{Example params confined}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 23 columns.}
\usage{
example_params_confined
}
\description{
Example params confined
}
\keyword{internal}
