% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\docType{data}
\name{example_params_unconfined}
\alias{example_params_unconfined}
\title{Example params unconfined}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 24 columns.}
\usage{
example_params_unconfined
}
\description{
Example params unconfined
}
\keyword{internal}
