% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_evaluate.R
\name{gather_outcomes}
\alias{gather_outcomes}
\title{Gather outcomes}
\usage{
gather_outcomes(treaty_df)
}
\arguments{
\item{treaty_df}{Outcomes from \code{evaluate_treaty_cases}, including utility, depth, pumping}
}
\description{
Gather outcomes
}
\examples{
library(genevoisgame)
library(ggplot2)
library(tidyr)
params <- example_params_confined
params$gs <- NULL
params <- crossing(params,gs=seq(0,1,by=0.05))
treaty_df <- evaluate_treaty_cases(params,'qudp')
treaty_long <- gather_outcomes(treaty_df)

ggplot(treaty_long) +
  geom_line(aes(x=gs,y=val,color=country,linetype=variable_subcat)) +
  facet_wrap(~variable_cat,scales="free_y",ncol=1) +
  scale_linetype_manual(values=c("solid","dashed","dotted","longdash")) +
  theme(legend.key.width = unit(1,"cm"))
}
