% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_evaluate.R
\name{get_contours}
\alias{get_contours}
\title{Get contour lines}
\usage{
get_contours(df = NULL, levels = 0, ...)
}
\arguments{
\item{df}{A data.frame containing x, y, and z columns. x and y must form a raster,
meaning every x must be represented for each y, and vice versa.}

\item{levels}{A vector of values at which contours should be plotted. If used, nlevels is ignored.}

\item{...}{If x, y, and z are directly specified here, df will be ignored.}
}
\description{
Get contour lines, using a wrapper for contourLines
}
\examples{
df <- tidyr::crossing(x=-10:10,y=-10:10) \%>\% dplyr::mutate(z=x^2)
cl <- get_contours(df,levels=5)
unique(cl$level)
ggplot() +
  geom_raster(data=df,aes(x,y,fill=z)) +
  geom_path(data=cl,aes(x,y,group=line))

cl <- get_contours(df,levels=c(15,20,60))
unique(cl$level)

df <- tidyr::crossing(x=seq(-5,5,length.out=20),y=seq(-5,5,length.out=20)) \%>\% dplyr::mutate(z=sqrt(x^2+y^2))
cl <- get_contours(df,levels=seq(2,10,by=2))
ggplot() +
  geom_raster(data=df,aes(x,y,fill=z)) +
  geom_path(data=cl,aes(x,y,group=line)) + coord_equal()
}
