% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_helpers.R
\name{in_range}
\alias{in_range}
\title{Check if value is within an inclusive range}
\usage{
in_range(x, range)
}
\description{
Check if value is within an inclusive range
}
\examples{
in_range(c(-3,-1,0,5,10,14),c(0,10))
}
