% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_params.R
\name{check_params}
\alias{check_params}
\title{Check for valid parameters}
\usage{
check_params(params)
}
\arguments{
\item{params}{\code{data.frame} or \code{list} of parameters}
}
\value{
Returns the aquifer type, depending on whether Dxx is specified ("confined") or PHIxx is specified ("unconfined")
}
\description{
Check for valid parameters
}
\details{
To evaluate the game, a \code{data.frame} or \code{list} of paramers must be supplied to
\code{evaluate_treaty} or \code{evaluate_treaty_cases}. These must include, for all scenarios:
\describe{
\item{Demand}{\code{Qf, Qs}}
\item{Unit cost of alternative supply}{\code{p0f, p0s}}
\item{Unit cost of pumping}{\code{B}}
\item{Unit cost of recharge, Swiss}{\code{crs}}
\item{Volume of recharge, with (T) and without (N) a treaty}{\code{rmT, rmN}}
\item{Trust between players}{\code{gs, gf}}
\item{Cost of the treaty}{\code{es, ef}}
}
And additional parameters depending on the type of aquifer. For confined aquifers:
\describe{
\item{Drawdown relationships}{\code{Dff, Dss, Dsf, Dfs}}
\item{Groundwater depth without pumping}{\code{d0s, d0f}}
\item{Recharge relationship with a treaty}{\code{DrsT, DrfT}}
\item{Recharge relationship without a treaty}{\code{DrsN, DrfN}}
}
And for unconfined aquifers:
\describe{
\item{Drawdown relationships}{\code{PHIff, PHIss, PHIsf, PHIfs}}
\item{Groundwater depth without pumping}{\code{dBs, dBf, h0s, h0f}}
\item{Recharge relationship with a treaty}{\code{PHIrsT, PHIrfT}}
\item{Recharge relationship without a treaty}{\code{PHIrsN, PHIrfN}}
}
}
\examples{
library(genevoisgame)
check_params(example_params_confined)
check_params(example_params_unconfined)
check_params(example_params_confined \%>\% dplyr::select())

params <- example_params_confined \%>\% dplyr::select(-gs) \%>\% tidyr::crossing(gs=c(0.2,0.8))
check_params(params)
check_params(params \%>\% dplyr::select())
}
