% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_helpers.R
\name{MM2R}
\alias{MM2R}
\title{Convert Mathematica to R}
\usage{
MM2R(path, ...)
}
\arguments{
\item{path}{Path to file to convert}

\item{...}{Arguments to change, listed in vector format, e.g.: \code{c(" ","*"))}}
}
\value{
Returns the corrected equations
}
\description{
Convert Mathematica to R
}
\examples{
eqns <- MM2R(path="mathematica/unconfined_eqns.txt", g1=c(" ","*"), g1=c("Sqrt","sqrt"))
}
